refactor courses endpoints
updated courses endpoints to reflect new db schema
