+++ server/endpoints/courses/courses_endpoints.ts
@@ -1,55 +1,15 @@
 import { courses } from '../../server';
 import { Course, courseContent, rosterSem } from '../../types';
 
+
+const currSem = 'SP22';
 export async function getCourses(req, res) {
   const courseId = req.query.id;
   const courseCode = req.query.code;
   const localCourses: Course[] = [];
-  let collectionIncrementer = 0;
 
   if (courseId === undefined) {
-    const courseTypes = await courses.doc(rosterSem).listCollections();
-    const collections = courseTypes.map((collection) => collection.listDocuments());
-    /* eslint-disable no-await-in-loop */
-    for (const collection of collections) {
-      const cId = courseTypes[collectionIncrementer].id;
-      for (const docRef of await collection) {
-        const cCode = docRef.id;
-        const cContent: courseContent = (await docRef.get()).data() as courseContent;
-        const course: Course = {
-          id: cId,
-          code: parseInt(cCode, 10),
-          content: cContent,
-        };
-        localCourses.push(course);
-      }
-      collectionIncrementer += 1;
-    }
-  } else {
-    const desiredCourse = (
-      await courses.doc(rosterSem).collection(courseId.toString()).get()
-    ).docs.filter((doc) => doc.id == courseCode.toString());
-    for (const doc of desiredCourse) {
-      const cContent: courseContent = doc.data() as courseContent;
-      const course: Course = {
-        id: courseId.toString(),
-        code: parseInt(courseCode.toString(), 10),
-        content: cContent,
-      };
-      localCourses.push(course);
-    }
-  }
-  res.send({ success: true, data: localCourses });
-}
-
-export async function getCoursesv2(req, res) {
-  const courseId = req.query.id;
-  const courseCode = req.query.code;
-  const localCourses: Course[] = [];
-  const collectionIncrementer = 0;
-
-  if (courseId === undefined) {
-    const coursesCollection = (await courses.doc('test').collection('test').get()).docs;
+    const coursesCollection = (await courses.doc(currSem).collection(currSem).get()).docs;
     /* eslint-disable no-await-in-loop */
     for (const doc of coursesCollection) {
       const docId = doc.id.split(' ');
@@ -64,7 +24,7 @@ export async function getCoursesv2(req, res) {
       localCourses.push(course);
     }
   } else {
-    const desiredCourse = (await courses.doc('test').collection('test').get()).docs.filter(
+    const desiredCourse = (await courses.doc(currSem).collection(currSem).get()).docs.filter(
       (doc) => doc.id == `${courseId} ${courseCode.toString()}`
     );
     for (const doc of desiredCourse) {
@@ -98,8 +58,8 @@ export async function createCourses(req, res) {
   ) {
     res.send({ success: false, message: 'One or more fields is missing.' });
   } else {
-    const courseIdCollection = courses.doc(rosterSem).collection(courseId);
-    const newCourse = courseIdCollection.doc(courseCode.toString());
+    const courseCollection = courses.doc(currSem).collection(currSem);
+    const newCourse = courseCollection.doc(courseId + ' ' + courseCode.toString());
     newCourse.set(course.content);
     res.send({ success: true, data: course });
   }
@@ -112,7 +72,7 @@ export async function deleteCourses(req, res) {
   if (courseId === undefined || courseCode === undefined) {
     res.send({ success: false, message: 'Course ID or code is missing.' });
   } else {
-    courses.doc(rosterSem).collection(courseId).doc(courseCode.toString()).delete();
+    courses.doc(currSem).collection(currSem).doc(courseId + ' ' + courseCode.toString()).delete();
     res.send({ success: true });
   }
 }
@@ -132,9 +92,9 @@ export async function updateCourses(req, res) {
     res.send({ success: true, message: 'One or more fields is missing.' });
   } else {
     courses
-      .doc(rosterSem)
-      .collection(courseId)
-      .doc(courseCode.toString())
+      .doc(currSem)
+      .collection(currSem)
+      .doc(courseId + ' ' + courseCode.toString())
       .update({ field: content });
     res.send({ success: true });
   }
